Update Song.java
Made the scanner read after the "time signature" on the db.txt file. I changed txt file so that after each measure, there are a series of numbers that tell us which beats are accented. The scanner now reads these and stores them in a new boolean array. The array starts as all false booleans with a length of the measure. When the scanner reads the numbers after the time signature, it replaces those booleans with true (so a 1 replaces index 0 with true). This array is later used in the Play class to separate accented and non-accented beats